Add sorting and filtering options to TagParams
The provided update expands the functionality of the TagParams class to allow for more detailed sorting and filtering. Sorting can now be done based on creation time and type. Additionally, the '@JsonIgnore' annotation was used to prevent the 'getFilterDefaults' method from being included in JSON serialization, preserving the default filter settings.