Refactor services and repositories to use new CrudServiceRepository pattern
Refactored the services and repositories interfaces to adjust to the new CrudServiceRepository design pattern. This introduces getToDtoMapper and getToEntityMapper in the repository layer and eliminates some redundant code. Furthermore, fixed the filename typo in PermissionToJpaMapper and updated Constants to include RECYCLONE_JPA_DEFAULT2. This is aimed at enhancing consistency and code maintainability.